@@ -1,16 +1,13 @@
 import { useCallback, useEffect, useRef, useState } from "react";
 import { useLocalStorage } from "usehooks-ts";
-import { Hex, HttpTransport, PrivateKeyAccount, createWalletClient, http } from "viem";
+import { Chain, Hex, HttpTransport, PrivateKeyAccount, createWalletClient, http } from "viem";
 import { generatePrivateKey, privateKeyToAccount } from "viem/accounts";
-import { Chain, WalletClient, usePublicClient } from "wagmi";
+import { WalletClient, usePublicClient } from "wagmi";
 
 const burnerStorageKey = "scaffoldEth2.burnerWallet.sk";
 
 /**
- * Is the private key valid
- * @internal
- * @param pk
- * @returns
+ * Checks if the private key is valid
  */
 const isValidSk = (pk: Hex | string | undefined | null): boolean => {
   return pk?.length === 64 || pk?.length === 66;
@@ -19,13 +16,10 @@ const isValidSk = (pk: Hex | string | undefined | null): boolean => {
 /**
  * If no burner is found in localstorage, we will generate a random private key
  */
-const newDefaultPriaveKey = generatePrivateKey();
+const newDefaultPrivateKey = generatePrivateKey();
 
 /**
- * Save the current burner private key from storage
- * Can be used outside of react.  Used by the burnerConnector.
- * @internal
- * @returns
+ * Save the current burner private key to local storage
  */
 export const saveBurnerSK = (privateKey: Hex): void => {
   if (typeof window != "undefined" && window != null) {
@@ -34,10 +28,7 @@ export const saveBurnerSK = (privateKey: Hex): void => {
 };
 
 /**
- * Gets the current burner private key from storage
- * Can be used outside of react.  Used by the burnerConnector.
- * @internal
- * @returns
+ * Gets the current burner private key from local storage
  */
 export const loadBurnerSK = (): Hex => {
   let currentSk: Hex = "0x";
@@ -48,47 +39,27 @@ export const loadBurnerSK = (): Hex => {
   if (!!currentSk && isValidSk(currentSk)) {
     return currentSk;
   } else {
-    saveBurnerSK(newDefaultPriaveKey);
-    return newDefaultPriaveKey;
+    saveBurnerSK(newDefaultPrivateKey);
+    return newDefaultPrivateKey;
   }
 };
 
-/**
- * #### Summary
- * Return type of useBurnerSigner:
- *
- * ##### ✏️ Notes
- * - provides signer
- * - methods of interacting with burner signer
- * - methods to save and loadd signer from local storage
- *
- * @category Hooks
- */
-export type TBurnerSigner = {
+type BurnerAccount = {
   walletClient: WalletClient | undefined;
   account: PrivateKeyAccount | undefined;
-  /**
-   * create a new burner signer
-   */
+  // creates a new burner account
   generateNewBurner: () => void;
-  /**
-   * explictly save burner to storage
-   */
+  // explicitly save burner to storage
   saveBurner: () => void;
 };
 
 /**
- * #### Summary
- * A hook that creates a burner signer/address and provides ways of interacting with
- * and updating the signer
- *
- * @category Hooks
- *
- * @param localProvider localhost provider
- * @returns IBurnerSigner
+ * Creates a burner wallet
  */
-export const useBurnerWallet = (): TBurnerSigner => {
-  const [burnerSk, setBurnerSk] = useLocalStorage<Hex>(burnerStorageKey, newDefaultPriaveKey);
+export const useBurnerWallet = (): BurnerAccount => {
+  const [burnerSk, setBurnerSk] = useLocalStorage<Hex>(burnerStorageKey, newDefaultPrivateKey, {
+    initializeWithValue: false,
+  });
 
   const publicClient = usePublicClient();
   const [walletClient, setWalletClient] = useState<WalletClient<HttpTransport, Chain, PrivateKeyAccount>>();
@@ -96,16 +67,10 @@ export const useBurnerWallet = (): TBurnerSigner => {
   const [account, setAccount] = useState<PrivateKeyAccount>();
   const isCreatingNewBurnerRef = useRef(false);
 
-  /**
-   * callback to save current wallet sk
-   */
   const saveBurner = useCallback(() => {
     setBurnerSk(generatedPrivateKey);
   }, [setBurnerSk, generatedPrivateKey]);
 
-  /**
-   * create a new burnerkey
-   */
   const generateNewBurner = useCallback(() => {
     if (publicClient && !isCreatingNewBurnerRef.current) {
       console.log("🔑 Create new burner wallet...");
@@ -125,7 +90,7 @@ export const useBurnerWallet = (): TBurnerSigner => {
       setAccount(randomAccount);
 
       setBurnerSk(() => {
-        console.log("🔥 ...Save new burner wallet");
+        console.log("🔥 Saving new burner wallet");
         isCreatingNewBurnerRef.current = false;
         return randomPrivateKey;
       });
